VISTARA : MailBox command implementation for DDR SPD error counter.
Summary:
In this task, mailbox command to store and clear DDR SPD error is implemented.
here 2 types of SPD errors are handled . Type 0 is SPD CRC ERROR and TYPE 1 is SPD NULL DATA ERROR
TYPE 0(SPD CRC) ERROR is checked for all 4 DIMMs and if error is found, when comparing calculated CRC with available CRC, then DIMM_ID and the offset
is stored in NVDATA
TYPE 1(SPD NULL DATA) ERROR is checked for all 4 DIMMs and if error is found,when reading first 4 bytes of the reading address, then DIMM_ID and the offset
is stored in NVDATA
MailBox command ddr-err-spd-info-get(0xFB38) is implemented to retrive this error information from persistent memory (NVDATA) and send as a response to this command.
MailBox command ddr-err-spd-info-clr(0xFB39) is implemented to clear this error information from NVDATA (persistent memory)
code cleanup for ddr-err-bist-info-clr(0xFB37) and ddr-err-bist-info-get(0xFB36) is implemented

Test Plan:
1. Modify the code to generate SPD NULL DATA error and if found then store the information in NVDATA.
2. Retrive the error information using ./cxl ddr-err-spd-info-get mem0/mem1
3. Reboot the host and again check if the information is persist
4. Clear this error information using ./cxl ddr-err-spd-info-clr mem0/mem1 -i dimm_id
5. Retrive the error information using ./cxl ddr-err-spd-info-get mem0/mem1
6. Reboot the host and again check if the information is persist

Reviewers:

Subscribers:

Tasks:

Tags: